Factor out static methods from mixmatch
